software: Add drivers, headers, and software tests
Consolidates C drivers and public headers for all peripherals and
system IPs added in this series. Adds a software test suite with
per-IP test cases and a CI workflow.

+++ software/driver/crc32.c
@@ -6,7 +6,7 @@
 
 #include "crc32.h"
 
-int crc32_init_driver(struct crc32_driver *driver, unsigned int base_address)
+int crc32_init_driver(struct crc32_driver *driver, unsigned long base_address)
 {
 	driver->regs = (volatile struct crc32_regs *)base_address;
