+++ software/include/syscon.h
@@ -69,7 +69,7 @@ struct syscon_driver {
 	volatile struct syscon_regs *regs;
 };
 
-void syscon_init(struct syscon_driver *driver, unsigned int base_address);
+void syscon_init(struct syscon_driver *driver, unsigned long base_address);
 
 unsigned int syscon_vendor(struct syscon_driver *driver);
 unsigned int syscon_platform(struct syscon_driver *driver);
